requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.5
django-cors-headers==4.4.0
django_csp==3.8
djangorestframework==3.15.2
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
